createVerificationVU implementation fix
