+++ src/src/presentation/dashboard/vuIdentity/IdentityScreen.tsx
@@ -7,9 +7,6 @@ import { DidiText } from "../../util/DidiText";
 import { DashboardScreenProps } from "../home/Dashboard";
 
 import strings from "../../resources/strings";
-import { didiConnect } from '../../../store/store';
-import { ActiveDid } from '../../../store/reducers/didReducer';
-import { createVerificationVU } from '../../../services/vuSecurity/createVerification';
 
 export interface IdentityScreenNavigation {
 	DashboardHome: DashboardScreenProps;
@@ -24,27 +21,12 @@ interface ItemsBtn {
 	navigation: ()=>any,
 }
 
-interface IdentityProps {
-    did: ActiveDid,
-    name: string,
-    lastname:string,
-}
-interface IdentityScreenDispatchProps {
-	vuSecurityData : (operationId: number, userName: string) => void;
-}
 
-export type IdentityScreenProps = IdentityProps &  IdentityScreenDispatchProps;
-
-class IdentityScreen extends NavigationEnabledComponent<IdentityScreenProps, {}, IdentityScreenNavigation> {
+export class IdentityScreen extends NavigationEnabledComponent<{}, {}, IdentityScreenNavigation> {
 	static navigationOptions = NavigationHeaderStyle.withTitleAndFakeBackButton<
 		IdentityScreenNavigation,
 		"DashboardHome"
 	>(strings.tabNames.identity, "DashboardHome", {});
-	async componentDidMount(){
-		const {did,lastname,name} = this.props;
-		const result = await createVerificationVU(did,name, lastname);
-		this.props.vuSecurityData(result.data.operationId,result.data.userName);
-	}
 
 	identityButtons: ItemsBtn[]  = [
 		{
@@ -95,21 +77,6 @@ class IdentityScreen extends NavigationEnabledComponent<IdentityScreenProps, {},
 		);
 	}
 }
-
-const connected = didiConnect(
-	IdentityScreen,
-	(state): IdentityProps => ({
-		did: state.did.activeDid,
-		name: state.persistedPersonalData.name,
-		lastname: state.persistedPersonalData.lastname
-}),
-(dispatch): IdentityScreenDispatchProps => ({
-	vuSecurityData : (operationId: number, userName: string) =>
-	dispatch({ type: "VU_SECURITY_DATA_SET", state: {operationId:`${operationId}`, userName } }),	
-})
-);
-
-export { connected as IdentityScreen };
 
 const styles = StyleSheet.create({
 	body: {
